refactor: extract github app common helpers
